UML Classes Created
The UML diagram for the Bear's Journey game, created in draw.io, outlines the overall classes for this project.

Some of these classes have been created in the Unity project but are not yet integrated into the game.

Classes like Goat and Sheep should be incorporated, taking into account their relationship with the Creature class, for example.